Apis of Owner and Admin

+++ Salon_Mgmt_Backend/src/main/java/com/salon/services/SalonServices.java
@@ -10,11 +10,11 @@ import org.springframework.web.multipart.MultipartFile;
 import com.salon.entities.Salon;
 
 public interface SalonServices {
-  public Salon createSalon(Long ownerId,SalonRequest req);
-  public Salon updateSalon(Long salonId, Long ownerId, SalonRequest req);
-  public Salon uploadLogo(Long salonId, MultipartFile file, Long ownerId) throws IOException;
-  public Salon uploadGallery(Long salonId, List<MultipartFile> files, Long ownerId) throws IOException ;
+  public Salon createSalon(Integer ownerId,SalonRequest req);
+  public Salon updateSalon(Long salonId, Integer ownerId, SalonRequest req);
+  public Salon uploadLogo(Long salonId, MultipartFile file, Integer ownerId) throws IOException;
+  public Salon uploadGallery(Long salonId, List<MultipartFile> files, Integer ownerId) throws IOException ;
   
-  public Salon getMySalon(Long salonId, Long ownerId);
-  public Salon submitSalon(Long salonId, Long ownerId);
+  public Salon getMySalon(Long salonId, Integer ownerId);
+  public Salon submitSalon(Long salonId, Integer ownerId);
 }
